Rename prop-types file
